+++ app/(tabs)/index.tsx
@@ -1,10 +1,11 @@
 import { Text, View, StyleSheet } from "react-native";
 import { Link } from 'expo-router';
+import Login from "./login";
 
 export default function Index() {
   return (
     <View style={styles.container}>
-      <Text style={styles.text}>Home Screen</Text>
+      <Login />
     </View>
   );
 }
